Make autosave optional for accepts_nested_attributes_for
accepts_nested_attributes_for always sets autosave to true on the
nested association. This will trigger the autosave code for the association
even if the association hasn't been changed. This becomes especially
problematic if the autosave has been set bi-directionally. If a child
object is saved it triggers the autosave on the parent which triggers
the autosave on the child again.

Instead we can only autosave if the nested_attributes have been set and
otherwise ignore them. This also prevents the bi-directional autosave in
most cases.