Clean up checks to see if DidYouMean.correct_error is defined
As of Ruby 2.7 DidYouMean is included as a default gem, so there is no need
to check if it is defined in the test suite.
Someone might still run Rails with DidYouMean disabled by using the
`--disable-did_you_mean` flag.

Since the included version of DidYouMean (1.4.0) also supports
`DidYouMean.correct_error`, we no longer need to check if the method is
defined.
See:
https://stdgems.org/2.7.0/
https://github.com/ruby/did_you_mean/blob/v1.4.0/lib/did_you_mean.rb#L89-L93

This also includes some of the changes made by Yuki in:
https://github.com/rails/rails/pull/39555

Co-Authored-By: Yuki Nishijima <yk.nishijima@gmail.com>